Use Evt instead of ee-ts
This is to support a Deno port in the near future, and should also
increase performance as ee-ts is known to have some performance issues.

Evt is not a drop-in replacement, which resulted in some refactoring. It
does support subclassing and using a single emitter bus, but the API for
doing this is pretty unattractive. Instead, its docs recommends to
compose many Evt emitters each suited to a single type of message, which
is what we do now.

Works toward #24