Bump sdk-js version
This isn't a breaking change in most usage, the public API of
`createIntervalHost` hasn't changed, just its internals. This is
backward compatible with a newer version as the Interval host.

While technically someone could be importing from
`interval-js/dist/ISocket` directly, we aren't exporting that interface
from the package entrypoint, so I think it's a fine policy to do this
without bumping major version.