Change spreadsheet columns type to a record, fix type deduction
Changing it to a record instead of an array of strings or definitions
makes it possible to deduce statically in IDEs what the return type of
the call will be (as well as improving the API I think).

Type narrowing is not yet performed when the spreadsheet call is part of
a renderGroup.