Explicitly mark which methods resolve immediately
This API also isn't amazing, but it feels better than implicitly
returning all `DISPLAY_` methods immediately. Having to do a property
existence check with `in` is less ideal than checking for truthiness,
but TypeScript doesn't like checking for truthiness when the property
doesn't exist for all schema methods. Despite that, I still like storing
this information directly in the schema as a source of truth to avoid
having to look elsewhere to know if something is immediate or not.
If that becomes a problem we can change the checker function definition.

Closes #103