Restructure createIntervalHost to an Interval class architecture
This should allow us to handle more advanced flows, like supporting
queues as mentioned in #91.

JavaScript doesn't have the prettiest class syntax, in particular there
are now a lot of `#`s strewn about for private properties and methods.
We could refactor `listen` to a single method with internal local
bindings to most of the variables and not actually bend them to the
Interval class, but this feels more correct despite looking busy.