Add queued actions
Actions can be queued from the host server using `interval.actions.enqueue()`
and dequeued using `interval.actions.dequeue()`.

Queued actions are displayed at the top of the organization dashboard.
Users see queued actions that are either assigned to no one or assigned
to them directly. When organization groups are added they should be able
to be assigned to groups as well, this currently only supports assigning
to users by email address or to anyone in the organization.

To not clobber any potential user-set search params on the new
transaction page, the queued action ID is passed to the newly-created
transaction via history state. We may want to consider reserving some
query params for our own use, or think of another solution.

Once a queued action has been started (a transaction for it exists)
it will only be visible to the owner of the transaction, unless
cancelled.