Slightly improve error message for grouping exclusive components
This doesn't yet actually prevent them from doing this at runtime
(we could do that as well), but adds an otherwise useless marker
property to groupable IOPromises that we erase for components that
should be rendered exclusively. We also now use the ioSchema directly as
the source of truth for this instead of maintaining a separate blocklist
of method names elsewhere.