Send sdk name and version when initializing host
This is very straightforward, just nullable strings, in order to be the
most flexible for any future SDKs and versions we may have.

The actual functional changes are very minor, but it required a
workaround to once again get TypeScript to not build files to
`dist/src`, because `package.json` is outside of the project root (`src`).

This is slightly hacky, but works because TypeScript compiles imports to
`require` statements,  and `../package.json` resolves after being built.

Closes #238