Change findAndSelect input signature
We now require `onSearch`, and `options` is renamed to `initialOptions`
and is now optional.

This ended up being relatively straightforward to override in the
constructor wrapper, and allows us to not need a fully new
`FIND_AND_SELECT` ioSchema method. The types ensure that our wrapper
passes through the correct values to the underlying `SELECT_SINGLE`
call, so it's still type safe.

Closes #187