Actually make running queued actions work in console
This "reuses" the queued action if the user presses "New transaction",
to allow for testing multiple runs of queued actions if desired.
After going back to the console, the queued action is no longer visible
because it has been completed.

We show a subtitle in the transaction header signifying that a queued
action is linked with the current transaction.

Closes #266