Allow components to modify their output, use for jsDate
I used a similar `get_value` function in the IOPromise classes in the
Python SDK and they worked well, this will allow further customization
of output and decouple them a bit from the defined schemas.

It's really more correct to call the schema values `response` instead of
`returns`, but I don't think that's worth changing at this time.

Closes #297