Use type variable instead of argument for output type narrowing
Using a type argument works sometimes, it's both unintuitive
(uses `extends` for assignment) and breaks down when used inline
sometimes. Using a type variable is clearer, and gives TypeScript a
point at which it can cement the value of the type, making it clear that
the output depends on the props input and not vice versa.