Handle undefined/null labels/values in table data
This allows `null` or `undefined` values to be used as values in table
rows, removing the need for developers to manually assign a default
empty value if they don't care. Note that we must cast it to null if
it's actually undefined before serializing it, because `undefined`
properties are completely stripped with `JSON.stringify`, and we need
the keys to be present after deserializing.

Closes #354