Add io.display.link and action-based links to IO tables
`io.display.link` is a new IO method that does what it says on the tin:
it renders the given URL as an `IVButton` button link.

In addition to the existing optional `href` property in the
display/select table column renderers, optional `action` and
`params` properties are now available. If `action` and optionally
`params` are provided and `href` is not, then an internal link
will be generated that links to the action of the provided slug.