Refactor JS SDK a bit, hopefully clarifying how things work together
Move many helper types into a new module. There are quite a few of these
currently, which made the file in which they were defined seem a lot
more complicated than it actually was. Abstracting these into their own
module reduces this burden a bit, since their names are relatively
self-descriptive and their implementations are somewhat unimportant to
the callers.

Replace several stateful functions with object returns to actual
classes (`IOClient`, `DuplexRPCClient`, `IOComponent`).
I personally was often confused about where things were defined, because
many core pieces where somewhat "magical" functions that maintained
their own state and returned an object. These classes behave in the
exact same way, but the simple act of actually calling them classes and
explicitly referring to class properties and methods instead of the deep
nested function scoping provides a lot of clarity to me, at least.

Refactor the custom component definition API. Custom components are
now straightforward functions that accept an input `Props` object and
return an object with the following optional properties: `props`, a
transformed version of its `Props` input which will be sent over the
network; `getValue`, a function that transforms the value received
from the network to the ultimate return value for the caller; and
`onStateChange`, a handler that receives the new state value from the
network and returns a new props object to send over the network.
This somewhat decouples the properties defined by the action developer
from the properties actually sent over the network, because they're
defined completely by the custom components and do not need to extend
the network types.
The type of this function is defined by `IOComponentDefinition` in the
new `types` module.

Add a README overview and brief JSDoc comments describing the
fundamental pieces of the SDK.