Add index and timestamp to SEND_LOG calls, sort by index
We need to add metadata on the host side to ensure accurate and
consistent log ordering in the console. We need to set the timestamps in
the host side in case the network calls are delayed due to flaky
transactions, and we need to also maintain an explicit index instead of
relying on the timestamp because multiple calls can be made within the
same millisecond.

This also adds a test to ensure that logs are sent and displayed in a
consistent order.

Closes #422