Add explicit action definitions and backgroundable property
In addition to simply specifying a callback for an action definition,
one can now also provide an explicit action definition object which
accepts a required `handler` property, and an optional `backgroundable`
boolean property.

In addition, when configuring an action's metadata, one is now able to
specify the `backgroundable` state for the action which if set will
override the code-defined property.

If `backgroundable` in the action metadata is set to "No" (`false`),
or if it is unset and is not `true` in the code definition, then any
in-progress transactions for that action will be automatically canceled
when the action runner leaves the transaction page.