Don't store last IO call in database, resend upon SDK reconnection
We still cache the last IO call in memory for faster response times on
reconnection (and a simpler architecture in the SDK), but no longer
persist it to storage and instead resend it from the SDK upon
reconnection for cases when the connection is dropped in the middle of a
transaction.

We unfortunately still need to persist the last IO call for older SDK
versions that do not perform this resending for now until support for
them is dropped.

Closes #442