Add ctx.loading API, remove io.experimental.progress methods
The loading state is sent and stored separately from IO calls.
If present, it displays a loading or progress indicator in the
Transaction UI, and is cleared when receiving a new IO call to render.

As discussed, `ctx.loading.start()` is used to start a new loading
state, `loading.update()` is used to update the loading state, merging
the current state with the supplied options, and
`loading.compeleteOne()` is used to mark a single queued item as
complete (for progress indicators).

Closes #325