More robustly handle unresponsive host connections
Don't immediately close the websocket connection when a ping fails,
instead mark as `UNRESPONSIVE` after a few minutes and continue to retry
for up to 6 hours before closing the connection and marking the host as
`OFFLINE`.

Additionally, also ping Interval from the host periodically, closing
the connection and reconnecting after a period of failed attempts
(3 minutes by default).