Show default indicator if no valid messages received after short time
We now show a default loading indicator if the action does not respond
with a valid render or loading call within a brief period, instead of
simply rendering a blank screen forever as before.

This can happen if the action is simply doing a lot of work, or if they
make invalid or malformed render calls that we don't know how to parse.

In these cases, we don't know exactly what went wrong to show a proper
error message, but we can at least show something that hints toward the
action not sending message properly.