Always use the HTTP endpoint for sending notifications
We have to keep the websocket endpoint around for now for backward
compatibility, but we're now logging when that occurs so we can
deprecate the feature at some point.

**Note**: This only currently works if the server that receives the HTTP
request is the same one that's hosting the websocket connection for the
transaction, otherwise relaying the notification to the connected client
will fail. Not sure how to handle this with our current architecture,
because I think we already don't fully support multiple websocket
servers correctly.