Add SdkAlert model, refactor INITIALIZE_HOST response
When a host is initialized, we now check its provided SDK name and
version against the `SdkAlert` records. If any alerts exist for that SDK
with a minimum supported version of greater than the one initializing,
then we return the alert along with the `INITIALIZE_HOST` response,
ordered descending by severity and version (they'll receive the most
severe alert, for the newest minimum supported version).

If any alerts are of severity `ERROR`, then the host initialization will
fail. If the connecting SDK supports receiving the new result `type`
field (of `success` or `error`) then it will receive the alert and an
`error` response. If it only supports `null` error responses, then we
must return `null` so the connection doesn't proceed.