Don't clear pendingIOCalls until transaction completed
We were mistakenly removing the pending call when it's successfully
resent, which means that the call would survive a single lost connection
to Interval, but if the connection is lost again it won't be able to
resend it again. This doesn't make any sense, we should only remove the
saved pending IO calls when the transaction has completed.