Add support for HttpHosts which dynamically create HostInstances
This supports not-always-on types of host workflows, primarily to
support serverless deployments, but supports any type of request.
This adds a `createHttpServer` helper that spins up a basic Node HTTP
server.

There's a bit of startup slowness involved with establishing the
connection, but that's not too bad for the time being I think.