Move persistent connection state into new IntervalClient class
This allows a lighter Interval class to exist as a user-facing frontend
for the IntervalClient class, along with the stateless HTTP API
functions that don't require a persistent connection to Interval.

This is probably for the best in some form regardless, the Interval
class was getting too big for its own good and had a large mix of
public and private internal state and functionality. Most of the private
internal state is now encapsulated into IntervalClient.

There's a little bit of duplication that should be removed in a future
slight rearchitecture (Interval.close just calling IntervalClient.close,
etc), but it's not too unwieldy.