Move configure page into a dedicated prefix route for clarity
It's possible to make the suffix still work, but it can get confusing
if we don't know that a given `/configure` is our own route or part of a
user slug. We can try to guess and probably usually get it right, but
it's probably best that we eliminate the possibility of getting it
wrong.