Strip defined-yet-undefined values from props before superjson
Superjson distinguishes defined values with the value of `undefined`,
so we strip them before passing them through. This doesn't really affect
anything in practice, since when deserializing they'll be recast to
`undefined` and then discarded, but this prevents them from being sent
over the wire at all.

This was really only a potential problem with a few of our custom
component overrides (io.search, etc), because we define props statically
but set them to undefined in order to convert the types, but could
potentially happen with conditional values in userland code as well.

Closes #642