Enforce io.select defaultValue being in provided options
If not the case, log a warning in the console and discard the provided
value (if `select.multiple`, discard the entire `defaultValue` array
instead of filtering).

To unify the `defaultValue` logic on our frontend, only use it when
determining the initial results array in TransactionUI, and use
the resulting `props.value` in each IO component's component.

Closes #669