Add .validate() functions to IOPromise and IOGroupPromise
This performs host-side validation on response values before actually
resolving the awaited IO call promise. Each validation function can
return either an error message string, or undefined if the response is
valid.

Because our current `io.group`s are array-based, it's awkward to specify
error messages for individual fields in the form-level validation. For
this reason, and because not all validation error messages necessarily
are associated with a single field but maybe just the form as a whole,
the group's validation function instead displays a single error message
at the bottom of the form if present.