Refactor ExclusiveIOPromise slightly, disallow optional
Because `io.confirm` is our only exclusive IO Promise right now, and
we don't want it to be able to be marked as `.optional()`, we now make
`ExclusiveIOPromise` its own distinct class separate from
`InputIOPromise`, identical except it cannot be optional.

This combination is mainly to avoid any further TypeScript overload
overload, it can be readded if/when we introduce any other Exclusive
methods.

Also, remove the unsightly `.exclusive()` method on `IOPromise` which
was unnecessary.

Closes #724