Add ctx.redirect() method
This currently "clears" the pending redirect as soon as it's sent to a
single client, in order to potentially facilitate returning to the
transaction afterward (maybe as part of an OAuth flow, for example).
That doesn't currently actually work very well for other tangential
reasons (no way to get url to current transaction, no way to patch
params in the middle of a transaction), but the redirect part should
facilitate it.

This means that reconnecting from another client will not redirect again
at this time. This behavior could be changed without a new SDK patch.