Fix `url` property for cell links
I think this has been broken forever, sadly. This simplifies the
`TableColumnResult` type so that TypeScript can now properly warn if you
specify an unknown value (`href`). We lose it knowing that `action` is
required for `params`, but the `href` thing is more confusing and
important to warn about I think.