Rename action slug prop to route
Using the `action` prop is deprecated, and is logged for `ctx.redirect`
and `io.display.link`. Like with the `href` prop in tables, we aren't
iterating through table data in order to log it for table cells or
menuItems. We may want to consider some way to do that.

I cannot get TypeScript to complain about extraneous keys in the
`renderCell` return type, so usages of `action` instead of `route`
in versions of the SDK moving forward will simply do nothing, but won't
trigger a TypeScript error alerting developers about it. Obviously not
ideal.

We haven't yet fully migrated from `actions` to `routes` in the
non-experimental API, and merging this will update the live docs for all
of the effected methods. Maybe we shouldn't do that yet? Or should not
make the documentation changes yet, and allow both properties for now?

Close T-338