Never clear result cache for io.search
It's probably not worth optimistically trying to save host memory
for cached searches in the same IO group, because it can lead to cases
where currently selected option is no longer present in the cache,
which causes a BAD_RESPONSE error when submitting the component.
The cached options point to the same underlying objects anyway, and
these shouldn't last that long, so it shouldn't be a problem to keep the
entire cache for the lifetime of the component.

Also add a test case to ensure two search components in the same group
don't break each other.

Fixes https://github.com/interval/interval-node/issues/14