Listen to iceconnectionstatechange to clean up client connection map
The `close` event in RTCDataChannel is just not being called for some
reason. It does not seem to be due to our code from what I can tell, we
are calling RTCDataChannel.close() but the event is not being fired.

Will want to investigate further before publicly releasing, but it is
behind a flag for now.