rename folder to `ui`, implement tables
We can call it whatever we want publicly, but "web-client" or "web-ui" is annoying for file lookups so `ui` gives it a short and unique namespace.

We're using `@ui/*` as the path alias in both `web` and `ui`. I briefly tried to get `~/*` working in each respective folder but AFAIK there's no way to tell TypeScript which is which.

I ended up needing to change a _lot_ of files here since a few core types and classes are now moved over to the UI package, but so far only tables and input.text are officially implemented.