Use/document more specific log methods (info, warn, etc), add "quiet"
Adds the new "quiet" logLevel option, which silences info and warn
messages (leaving only error and prod). The default logLevel is
now "info", which is an alias to the deprecated "prod".

Changes instances of Logger calls to be more mindful about their
importance. Generally means that many were downgraded
(prod -> info, error -> warn) if they're not fatal or a result of
errors in user code.

cc: https://github.com/interval/interval-node/issues/21