Perform clientId authentication in transactions and pages
Peer clients are authenticated through the Interval backend, ensuring
that they are authorized to connect on behalf of the organization,
however up until now we weren't validating individual transactions and
pages to ensure the requesting peer connection is the same client as
the initiator when connecting to in-progress transactions/pages.

We are already sending the `clientId` when creating transactions/pages
via the Interval backend, and these same IDs are used by the peer
connections as well, so we just need to check to make sure the `clientId`
is the same when someone tries opening an existing connection, treating
the (`transactionId`/`pageKey`, `clientId`) pair as a unique and trustworthy
authentication key. These are random and while the `transactionId` is
accessible by administrators in the transaction history, `clientId`s are
not.

The only legitimate flow this should affect is a peer-enabled user's
first transaction/page, where it is initiated via the Interval WSS
connection before being picked up by the peer connection that is
established immediately afterward.