Keep all IO response handlers around until transaction closed
The transaction is closed either at the completion of the handler as
before, or if `postponeCompleteCleanup` is set via the
`TRANSACTION_AUTO_CONTINUE` flag then when the new `CLOSE_TRANSACTION`
RPC call is sent (which behaves much like the existing `CLOSE_PAGE` call).

This allows stateful components to continue working after they have been
submitted, pretty much relevant for only stateful display components
like table and grid.

Some bugs remain around sometimes showing the continue button when we
shouldn't and sometimes attempting to re-submit an already-completed IO
call, because the frontend doesn't currently actually know whether an IO
call was submitted successfully or not until the subsequent call is sent
to us.