Add linear backoff and maximum retry logic for failed RPC calls
While it turns out that this wasn't causing issues in the wild yet, it's
possible to basically saturate a connection with very large payloads and
our existing naive and static retry logic. This introduces backoff with
a max number of retries, so one failed RPC call can't permanently
bring a host to its knees.