Use ws connection state and not our own flags
It seems like race conditions are somehow causing our isConnected flags
to be incorrect. Let's try removing those entirely and relying on the
socket's connection status directly instead.

We still use an isReconnecting flag to prevent the auto-reconnect logic
from happening multiple times, but this is only used for that and
shouldn't affect any actual calls or cause reconnect attempts themselves
to fail.